requirements.txt
Django==1.11.2
djangorestframework==3.6.3
django-cors-headers==2.1.0
